Replace loops with tensor operations
To achieve this, a lot of functions had to be changed so they work with
tensors that have more than 2 dimensions.

The changes mostly affected the functions in `inf_vi` that begin with
`_build`. But also the covariance and the likelihood had to be adapted.

One crucial component was a broadcasting `matmul` function. `tf.matmul`
does not broadcast which is a problem when you want to multiply two
tensors that don't have the same dimension. The current implementation
of a broadcasting `matmul` is based on `tf.einsum` which is a very
general tool for summing over certain dimensions in two tensors.