First version of the new main loop
The goal is to make it easy to run the code on a cluster while avoiding
having to write lots of boilerplate code.

The code is aiming to be configurable completely by setting flags so
that you don't have to change code when you try things out. This means,
the code is in some ways less flexible. But it should be enough for
doing research.

Several datasets have been added. Several test datasets with different
numbers of in and output dimensions. And the MNIST dataset.