Add a new training loop that uses eager execution
This is a first implementation which is not really optimized yet. Some
changes to the inference functions are needed to make it more efficient.

The training loop is mostly intended for debugging. As such it does not
contain code for creating summaries as this is usually not necessary
when debugging.