Restructure inference classes for eager execution
The main goal of this change is to be able to use exact inference in eager
execution. Another goal is to separate inference from prediction because
the code in the training loops was a bit hacky when both are done in the
same function.

Major changes:

* separate `predict` and `inference`
* move variable definition code out of `inference()` and into
`__init__()`
* reduce the number of parameters for the init function of
`inf.Variational` by putting extra parameters into a `params` structure
* get rid of `EeagerVariableStore`