Put metrics in separate file
The metrics for the graph training and the eager training are now in one
file `util/metrics.py`. Now it's possible to share code between these
two.

Furthermore, the metric code in `train_eager.py` was very weird. Now the
code is much cleaner.

Another change is, that metrics can now also depend on the *input* of
the test set. This will be useful when we want to have a metric that
measures fairness.