Offer to save predictions instead of variables
The values of the variables are easy to get from the saved model. The
predictions can actually be not so easy to get if the model is very
computationally expensive.

The predictions are stored as numpy files. It would have been very
tedious to save them as text files because then you have to indicate
somehow which column represents what.